+++ tsm-bouali.c
@@ -13,7 +13,7 @@
 
 typedef struct { mpfr_t a, b, g, m; series gx2; } parameters;
 
-void *get_p (int argc, char **argv, long n) {
+void *get_p (int argc, char **argv, int n) {
     assert(argc == 13);
     parameters *p = malloc(sizeof (parameters));
     t_params(argv, argc, &p->a, &p->b, &p->g, &p->m);
